fix(viz): unify ego/neighbor footprint wheelbase convention across viz + sim
Central library:
- Extract render_scene_at_step + draw helpers + render color constants into a
  new gradio-free scenario_generation/scene_render.py; scene_branch_editor and
  the new viz_saved_scenes tool import the renderer instead of duplicating it.

Convention (matches loss.py / reward.py: ego center = rear-axle + wheelbase/2,
span +/-length/2; neighbors are centroid-referenced):
- visualize.draw_trajectory: add wheelbase passthrough to footprint boxes.
- viz_p4_recovery, viz_prism_compare: ego anchor box ro = (length-wb)/2 (was
  length-wb, missing /2).
- viz_kinematic_floored, viz_t1_crossing_debug: center-shift by wb/2 (was
  (length-wb)/2).
- visualize_scenes, scene_render placed obstacles: neighbor boxes centroid
  (length/2), drop the length*0.65 rear-axle offset.
- recovery_sim._draw_agent_box: delegate to visualize.draw_agent_box with a
  wheelbase param; neighbor calls default to centroid, ego calls pass wheelbase.
- replay.py / simulate.py: ego plan-trajectory footprints pass wheelbase.

Correctness:
- advance_scene_mpc: add the neighbor centroid<->rear-axle conversion that
  advance_scene already does (inference frames a neighbor-as-ego at the rear
  axle); without it model-driven NPC positions were biased ~wheelbase/2 forward.
